Fixed particles almost

+++ src/gameplay.h
@@ -2,11 +2,12 @@
 #include "raylib.h"
 #include <stdlib.h>
 #include <math.h>
+#include <stdio.h>
 
 typedef struct Particle {
     Vector2 position;  // Current position
     // Vector2 velocity;  // Movement direction and speed
-    int velocity;
+    float velocity;
     float lifetime;    // Time before the particle disappears
     // float size;        // Size of the particle
     Color color;       // Particle color
